requirements.txt
discord.py
rHLDS
gspread
oauth2client
